separate confparse into own file
